biz-service/requirements.txt
Flask>=2.3.0
PyJWT>=2.8.0
redis>=4.5.0
